Adicionando números decimais

+++ Calculadora.py
@@ -1,6 +1,6 @@
-n1 = int(input("Digite o primeiro número: "))
+n1 = float(input("Digite o primeiro número: "))
 operacao = input("Digite a operação: ")
-n2 = int(input("Digite o segundo número: "))
+n2 = float(input("Digite o segundo número: "))
 
 
 match operacao:
